[9] Created prototype TestRunner class
