+++ index.js
@@ -3,11 +3,13 @@ const message = require("./src/message");
 const node = require("./src/node");
 const system = require("./src/system");
 const users = require("./src/users");
+const testRunner = require("./src/testRunner");
 
 module.exports = {
   ...field,
   ...message,
   ...node,
   ...system,
-  ...users
+  ...users,
+  ...testRunner
 };
